Revert all changes

+++ src/main/java/cn/nukkit/level/format/anvil/Chunk.java
@@ -275,7 +275,7 @@ public class Chunk extends BaseChunk {
             }
             CompoundTag s = new CompoundTag(null);
             s.putByte("Y", section.getY());
-            s.putIntArray("Blocks", section.getIdArray());
+            s.putByteArray("Blocks", section.getIdArray());
             s.putByteArray("Data", section.getDataArray());
             s.putByteArray("BlockLight", section.getLightArray());
             s.putByteArray("SkyLight", section.getSkyLightArray());
@@ -358,7 +358,7 @@ public class Chunk extends BaseChunk {
             }
             CompoundTag s = new CompoundTag(null);
             s.putByte("Y", (section.getY()));
-            s.putIntArray("Blocks", section.getIdArray());
+            s.putByteArray("Blocks", section.getIdArray());
             s.putByteArray("Data", section.getDataArray());
             s.putByteArray("BlockLight", section.getLightArray());
             s.putByteArray("SkyLight", section.getSkyLightArray());
